Updated provisioning script
Added python-software-properties to get add-repository (repo could be manually added to source-list, but that doesn't install the key as well; this is simpler)
Added the brightbox repo to get access to recent rubies
Added --force-yes to the apt-get command because otherwise it complains about it if it hits a problem. Ordinarily I'd say "never do this" (and I may yet back this one out) but this is a vagrant dev box so if it really does something nasty it's a simple thing to rebuild it, meanwhile it keeps the provisioning going so you can see what else is wrong.
Installed ruby 2.1 instead of 1.9 -- added the dev package to compile some extensions, the doc package because I'm anal that way. For the moment passing on the ruby tcl/tk package for now. Not sure if I'm going to need it.
Changed environment variables for ruby 2.1