Align README more with the current state of Radiant
The original README made some assumptions that don't seem in line with the current state of the app. I've removed commands that don't work (it defaults to sqlite as it is) and replaced the two rake db: tasks with the single app:db:bootstrap task, that's what Radiant currently uses to create *and* migrate in the same command. These changes are not permanent, there will be more as the app itself changes.